backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
cryptography
PyJWT
passlib
bcrypt==3.2.2
python-multipart